Added sprint function
There is currently an issue with the sprint function. If the player "sprints" to full speed then releases the sprint button they will still keep going at the same speed. Maybe not a huge issue but I tried to make the speed gradually lower to the normal max speed when not sprinting.